+++ tests/surfaces/test_revolutionsurface3d.py
@@ -17,21 +17,12 @@ class TestRevolutionSurface3D(unittest.TestCase):
     wire = vmw.Wire3D([linesegment, arc])
     axis_point = volmdlr.O3D
     axis = volmdlr.Z3D
-
-    def test_init(self):
-
-        surface = surfaces.RevolutionSurface3D(self.wire, self.axis_point, self.axis)
-
-        self.assertEqual(surface.x_periodicity, volmdlr.TWO_PI)
-        self.assertEqual(surface.y_periodicity, None)
-        self.assertTrue(surface.frame.origin.is_close(self.axis_point))
-        self.assertTrue(surface.axis_point.is_close(self.axis_point))
-        self.assertTrue(surface.axis.is_close(self.axis))
+    surface = surfaces.RevolutionSurface3D(wire, axis_point, axis)
 
     def test_point2d_to_3d(self):
         surface = surfaces.RevolutionSurface3D(self.wire, self.axis_point, self.axis)
 
-        point2d = volmdlr.Point2D(math.pi, 0.7047817224492219)
+        point2d = volmdlr.Point2D(math.pi, 0.5)
         point3d = surface.point2d_to_3d(point2d)
         expected_point3d = volmdlr.Point3D(-0.5, 0, 0.5)
 
@@ -42,7 +33,7 @@ class TestRevolutionSurface3D(unittest.TestCase):
 
         point3d = volmdlr.Point3D(-0.5, 0, 0.5)
         point2d = surface.point3d_to_2d(point3d)
-        expected_point2d = volmdlr.Point2D(math.pi, 0.7047817224492219)
+        expected_point2d = volmdlr.Point2D(math.pi, 0.5)
 
         self.assertTrue(point2d.is_close(expected_point2d))
 
@@ -51,6 +42,47 @@ class TestRevolutionSurface3D(unittest.TestCase):
         rectangular_cut = volmdlr.faces.RevolutionFace3D.from_surface_rectangular_cut(
             surface, 0, volmdlr.TWO_PI, 0, 1)
         self.assertEqual(rectangular_cut.surface2d.area(), volmdlr.TWO_PI)
+
+    def arc3d_to_2d(self):
+        surface = surfaces.RevolutionSurface3D.load_from_file(
+            "surfaces/objects_revolution_tests/revolution_surface_bug_0.json")
+        contour = vmw.Contour3D.load_from_file("surfaces/objects_revolution_tests/revolution_contour_bug_0.json")
+        arc = contour.primitives[3]
+        linesegment2d = surface.arc3d_to_2d(arc)[0]
+        self.assertTrue(linesegment2d.start.is_close(volmdlr.Point2D(-2.6665730021726306, 0.0003141532401719152)))
+        self.assertTrue(linesegment2d.end.is_close(volmdlr.Point2D(-2.6665730021726324, 0)))
+
+        surface = surfaces.RevolutionSurface3D.load_from_file(
+            "surfaces/objects_revolution_tests/revolution_surface_bug_1.json")
+        contour = vmw.Contour3D.load_from_file("surfaces/objects_revolution_tests/revolution_contour_bug_1.json")
+        arc = contour.primitives[3]
+        linesegment2d = surface.arc3d_to_2d(arc)[0]
+        self.assertTrue(linesegment2d.start.is_close(volmdlr.Point2D(1.0582040468439544, 0.0)))
+        self.assertTrue(linesegment2d.end.is_close(volmdlr.Point2D(6.0956438652626925, 0.0)))
+
+    def test_frame_mapping(self):
+        surface = self.surface
+        new_frame = volmdlr.Frame3D(volmdlr.Point3D(0, 1, 0), volmdlr.X3D, volmdlr.Y3D, volmdlr.Z3D)
+        new_surface = surface.frame_mapping(new_frame, "old")
+        self.assertEqual(new_surface.wire.primitives[0].start.y, 1)
+        self.assertTrue(new_surface.frame.origin.is_close(volmdlr.Point3D(0, 1, 0)))
+
+    def test_simplify(self):
+        rev1 = surfaces.RevolutionSurface3D.load_from_file(
+            "surfaces/objects_revolution_tests/revolutionsurface_simplify_spherical.json")
+        rev2 = surfaces.RevolutionSurface3D.load_from_file(
+            "surfaces/objects_revolution_tests/revolutionsurface_simplify_conical.json")
+        rev3 = surfaces.RevolutionSurface3D.load_from_file(
+            "surfaces/objects_revolution_tests/revolutionsurface_simplify_cylindrical.json")
+
+        sphere = rev1.simplify()
+        self.assertTrue(isinstance(sphere, surfaces.SphericalSurface3D))
+
+        cone = rev2.simplify()
+        self.assertTrue(isinstance(cone, surfaces.ConicalSurface3D))
+
+        cylinder = rev3.simplify()
+        self.assertTrue(isinstance(cylinder, surfaces.CylindricalSurface3D))
 
 
 if __name__ == '__main__':