+++ volmdlr/edges.py
@@ -2336,6 +2336,12 @@ class LineSegment2D(LineSegment):
         self._bounding_rectangle = None
         LineSegment.__init__(self, start, end, name=name)
 
+    def copy(self, deep=True, memo=None):
+        """
+        A specified copy of a LineSegment2D.
+        """
+        return self.__class__(start=self.start.copy(deep, memo), end=self.end.copy(deep, memo), name=self.name)
+
     def __hash__(self):
         # return self._data_hash()
         # tolerance = 1e-6
@@ -5591,6 +5597,12 @@ class LineSegment3D(LineSegment):
         # Cylindrical face
         return self._cylindrical_revolution([axis, u, p1_proj, distance_1, distance_2, angle])
 
+    def trim(self, point1: volmdlr.Point3D, point2: volmdlr.Point3D):
+        if not self.point_belongs(point1) or not self.point_belongs(point2):
+            raise ValueError('Point not on curve')
+
+        return LineSegment3D(point1, point2)
+
 
 class BSplineCurve3D(BSplineCurve):
     """
@@ -6105,6 +6117,18 @@ class BSplineCurve3D(BSplineCurve):
         if return_points:
             return minimum_distance, points[0], points[1]
         return minimum_distance
+
+    def frame_mapping(self, frame: volmdlr.Frame3D, side: str):
+        """
+        Returns a new Revolution Surface positionned in the specified frame.
+
+        :param frame: Frame of reference
+        :type frame: `volmdlr.Frame3D`
+        :param side: 'old' or 'new'
+        """
+        new_control_points = [control_point.frame_mapping(frame, side) for control_point in self.control_points]
+        return BSplineCurve3D(self.degree, new_control_points, self.knot_multiplicities, self.knots, self.weights,
+                              self.periodic, self.name)
 
 
 class BezierCurve3D(BSplineCurve3D):